Add draft for testing traccuracy implementation
