+++ scripts/test_ctc_metric.py
@@ -1,122 +1,72 @@
-import os
-from glob import glob
-from pathlib import Path
-
 import numpy as np
-import pandas as pd
-import imageio.v3 as imageio
 
-import napari
-
-from deepcell_tracking.utils import load_trks
-# from deepcell_tracking.isbi_utils import trk_to_isbi
+from deepcell_tracking.isbi_utils import trk_to_isbi
 
 from traccuracy.loaders._ctc import _get_node_attributes
 
-
-def load_tracking_segmentation(experiment):
-    ROOT = r"/home/anwai/results/tracking/MicroSAM testing/"
-    TRACKMATE_ROOT = r"/home/anwai/results/tracking/trackmate_stardist/microSAM revision every 3rd fr"
-
-    if experiment == "vit_l":
-        seg_path = glob(os.path.join(ROOT, r"round 2 vit_l", "*.tif"))[0]
-    elif experiment == "vit_l_lm":
-        seg_path = glob(os.path.join(ROOT, "vit_l_finetuned", "*.tif"))[0]
-    elif experiment == "vit_l_specialist":
-        seg_path = glob(os.path.join(ROOT, "vit_l_specialist", "*.tif"))[0]
-    elif experiment == "trackmate_stardist":
-        seg_path = glob(os.path.join(TRACKMATE_ROOT, "*.tif"))[0]
-    else:
-        raise ValueError(experiment)
-
-    return imageio.imread(seg_path)
+from get_tracking_results import get_tracking_data, load_tracking_segmentation
 
 
-def check_tracking_results(raw, labels, curr_lineages, chosen_frames):
-    seg_default = load_tracking_segmentation("vit_l")
-    seg_generalist = load_tracking_segmentation("vit_l_lm")
-    seg_specialist = load_tracking_segmentation("vit_l_specialist")
+def extract_df_from_segmentation(segmentation):
+    track_ids = np.unique(segmentation)[1:]
+    last_frame = segmentation.shape[0] - 1
 
-    # let's get the tracks only for the objects present per frame
-    for idx in np.unique(labels)[1:]:
-        lineage = curr_lineages[idx]
-        lineage["frames"] = [frame for frame in lineage["frames"] if frame in chosen_frames]
+    all_tracks = []
+    splits = 0
+    for idx in track_ids:
 
-    v = napari.Viewer()
-    v.add_image(raw)
-    v.add_labels(labels)
+        frames = np.unique(np.where(segmentation == idx)[0])
 
-    v.add_labels(seg_default, visible=False)
-    v.add_labels(seg_generalist, visible=False)
-    v.add_labels(seg_specialist, visible=False)
+        if frames.min() == 0:  # object starts at first frame
+            if frames.max() == last_frame:  # object is tracked until the last frame
+                pid = 0
+                have_fam = None  # they can't split in this case
+            else:  # object either goes out of frame or splits
+                pid = 0
+                have_fam = frames.max()  # let's assume that it splits, we will know if it does or not
 
-    napari.run()
+        else:
+            if have_fam is not None:  # takes the parent information from above
+                pid = have_fam
+                splits += 1
 
+                if splits > 2:  # assumes every mother cell splits into two daughter cells
+                    print("The mother cell has made enough daughter splits, hence this is a new object.")
+                    splits = 0
+                    # pid = 0  # this is the case where an objects appears at nth frame and has no parent id
+            else:
+                pid = 0  # assumes that it was an object that started at a random frame
 
-def get_tracking_data():
-    data_dir = "/home/anwai/data/dynamicnuclearnet/DynamicNuclearNet-tracking-v1_0/"
-    data_source = np.load(os.path.join(data_dir, "data-source.npz"), allow_pickle=True)
+        track_dict = {
+            "Cell_ID": idx,
+            "Start": frames.min(),
+            "End": frames.max(),
+            "Parent_ID": pid,
+        }
 
-    fname = "test.trks"
-    track_file = os.path.join(data_dir, fname)
-    split_name = Path(track_file).stem
+        print(track_dict)
+        all_tracks.append(track_dict)
 
-    data = load_trks(track_file)
-
-    X = data["X"]
-    y = data["y"]
-    lineages = data["lineages"]
-
-    meta = pd.DataFrame(
-        data_source[split_name],
-        columns=["filename", "experiment", "pixel_size", "screening_passed", "time_step", "specimen"]
-    )
-    print(meta)
-
-    # let's convert the data to expected shape
-    X = X.squeeze(-1)
-    y = y.squeeze(-1)
-
-    # NOTE: chosen slice for the tracking user study.
-    _slice = 7
-    raw, labels = X[_slice, ...], y[_slice, ...]
-    curr_lineages = lineages[_slice]
-
-    # NOTE: let's get every third frame of data and see how it looks
-    chosen_frames = list(range(0, raw.shape[0], 3))
-    raw = np.stack([raw[frame] for frame in chosen_frames])
-    labels = np.stack([labels[frame] for frame in chosen_frames])
-
-    # let's create a value map
-    frmaps = {}
-    for i, frval in enumerate(chosen_frames):
-        frmaps[frval] = i
-
-    # let's remove frames which are not a part of our chosen frames
-    for k, v in curr_lineages.items():
-        curr_frames = v["frames"]
-        v["frames"] = [frmaps[frval] for frval in curr_frames if frval in chosen_frames]
-
-    return raw, labels, curr_lineages, chosen_frames
+        breakpoint()
 
 
 def evaluate_tracking(raw, labels, curr_lineages, chosen_frames, segmentation_method):
     seg = load_tracking_segmentation(segmentation_method)
 
+    # calcuates node attributes for each detection
     gt_df = _get_node_attributes(labels)
     seg_df = _get_node_attributes(seg)
 
+    # converts inputs to isbi-track format - the version expected as inputs in traccuracy
     output = trk_to_isbi(curr_lineages, path=None)
 
-    breakpoint()
+    df = extract_df_from_segmentation(seg)
 
 
 def main():
     raw, labels, curr_lineages, chosen_frames = get_tracking_data()
 
-    # check_tracking_results(raw, labels, curr_lineages, chosen_frames)
-
-    segmentation_method = "trackmate_stardist"
+    segmentation_method = "vit_l_specialist"
     evaluate_tracking(raw, labels, curr_lineages, chosen_frames, segmentation_method)
